09/16 3

+++ shop/WebContent/deleteMember.jsp
@@ -6,6 +6,13 @@
 
 	int memberNo = Integer.parseInt(request.getParameter("memberNo"));
 	
+	if(request.getParameter("memberNo") == null){
+		response.sendRedirect(request.getContextPath() + "/index.jsp");
+		return;
+	}
+	
 	MemberDao memberDao = new MemberDao();
 	memberDao.deleteMemberByKey(memberNo);
+	
+	response.sendRedirect(request.getContextPath()+"/admin/selectMemberList.jsp");
 %>
